// src/types/Heading.types.ts
export type HeadingTypes = {
  headingType: "1" | "2" | "3" | "4" | "5" | "6";
  headingText: string;
  extraClasses?: string;
  variant:
    | "display-1"
    | "display-2"
    | "display-3"
    | "headline-1"
    | "headline-2"
    | "headline-3"
    | "headline-4"
    | "headline-5"
    | "headline-6";
};